% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoJuncWithShareSites.R
\name{annoJuncWithShareSites}
\alias{annoJuncWithShareSites}
\title{Find annotated junction that share a same annotated splice site with a novel junction}
\usage{
annoJuncWithShareSites(gene_junc, index, docking_site)
}
\value{
coordinates of junctions concatenated by "|"
}
\description{
Find annotated junction that share a same annotated splice site with a novel junction
}
