% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateJunc_.R
\name{annotateJunc_}
\alias{annotateJunc_}
\title{Annotate junctions}
\usage{
annotateJunc_(
  merged_junc,
  gene_ref,
  ensembl_pos,
  protein_coding_gene,
  intron_coding,
  intron_noncoding,
  addition_annotation
)
}
\value{
data.frame One row in the the data.frame is annotations of a unique junction.
}
\description{
Annotate junctions
}
