% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateLeafCutterJunc.R
\name{annotateLeafCutterJunc}
\alias{annotateLeafCutterJunc}
\title{Annotate junctions extracted from RNAseq data and filter for junction outliers from LeafCutter analysis result}
\usage{
annotateLeafCutterJunc(
  rna_meta,
  chrom,
  tissues_leafcutter_pvals_file = NULL,
  max_LeafCutter_Pval = 0.05,
  min_nr_tissue = 1,
  output_prefix,
  reference = "Default",
  addition_annotation = list(snaptron = system.file(paste("extdata", "Reference",
    "snaptron_gtex_anno.txt", sep = .Platform$file.sep), package = "SpliPath"))
)
}
\arguments{
\item{rna_meta}{string RNAseq sample metadata file. The file should contain at least "SampleID", "SubjectID", "Group", "Tissue", and "Path" columns}

\item{chrom}{string The junctions in this chromosome will be merged among samples.}

\item{tissues_leafcutter_pvals_file}{list The file paths to the LeafCutterMD analysis P values of each tissue type. Each element is a LeafCutterMD file path named under the tissue type. It should also contain an "colnamesToSampleID" element, which is a vector of RNAseq SampleIDs named by the matching column names in the LeafCutterMD files. The tissues should be the same with those in the RNAseq sample metadata file.}

\item{max_LeafCutter_Pval}{numeric The threshold of LeafCutterMD P value of a novel junction.}

\item{min_nr_tissue}{numeric Require novel junction LeafCutterMD P value under the provided threshold in at least min_nr_tissue tissues.}

\item{output_prefix}{string Prefix of output files.}

\item{reference}{string The directory to genomic annotation files, which can be generated by prepareGenoRef function. Default to pre-made ensembl annotation version GRCh38.98.}

\item{addition_annotation}{list or NULL Optional. Name and file name of additional junction annotation source. Default: list(snaptron = system.file("extdata/Reference/snaptron_gtex_anno.txt", package = "SpliPath")). Whether the input junctions are the given junction list will be shown in the output file (in column "in.annotation_name").}
}
\description{
Annotate junctions extracted from RNAseq data and filter for junction outliers from LeafCutter analysis result
}
