% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateQueryJunc.R
\name{annotateQueryJunc}
\alias{annotateQueryJunc}
\title{Annotate query junctions}
\usage{
annotateQueryJunc(
  query_junc,
  output_prefix,
  reference = "Default",
  addition_annotation = list(snaptron = system.file(paste("extdata", "Reference",
    "snaptron_gtex_anno.txt", sep = .Platform$file.sep), package = "SpliPath"))
)
}
\arguments{
\item{query_junc}{string BED file containing the coordinates of query junctions to be annotated.}

\item{output_prefix}{string Prefix of output file. The annotations of the junctions will be written in a prefix_anno.txt.gz file.}

\item{reference}{string The directory to genomic annotation files, which can be generated by prepareGenoRef function. Default to pre-made ensembl annotation version GRCh38.98.}

\item{addition_annotation}{list or NULL Optional. Name and file name of additional junction annotation source. Default: list(snaptron = system.file("extdata/Reference/snaptron_gtex_anno.txt", package = "SpliPath")). Whether the input junctions are the given junction list will be shown in the output file (in column "in.annotation_name").}
}
\description{
Annotate query junctions
}
