% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browserData.R
\name{browserData}
\alias{browserData}
\title{Prepare data files for SpliPath data browser}
\usage{
browserData(
  paired_data,
  junc_anno_files,
  cs_qtl_files,
  tissues_leafcutter_pvals_file = NULL,
  junc_count_files = NULL,
  output_dir = "."
)
}
\arguments{
\item{paired_data}{logical "TRUE" for paired data mode or "FALSE" for unpaired data mode. If TRUE, arguments tissues_leafcutter_pvals_file, junc_count_files, rna_meta, and wgs_meta are required. The function matches the paired samples by 'SubjectID' in the metadata.}

\item{junc_anno_files}{vector A list of output files of annotateJunc function, which contain junction annotations.}

\item{cs_qtl_files}{string The output file of collapsedsQTL function. One row in the the data.frame is an annotated DNA variant that mapped to a novel junction in a subject's tissues.}

\item{tissues_leafcutter_pvals_file}{list The file paths to the LeafCutterMD analysis P values of each tissue type. Each element is a LeafCutterMD file path named under the tissue type. It should also contain an "colnamesToSampleID" element, which is a vector of RNAseq SampleIDs named by the matching column names in the LeafCutterMD files. The tissues should be the same with those in the RNAseq sample metadata file.}

\item{junc_count_files}{vector A list of output files of mergeJuncData funciton, which contain junction-sample read count table. The input file name list in the first four arguments should be matched.}

\item{output_dir}{string The path to data output directory.}
}
\description{
Make splice junction tables and csQTL tables for each gene.
}
