% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calNormValue.R
\name{calNormValue}
\alias{calNormValue}
\title{Calculate PSI}
\usage{
calNormValue(gene_junc, index, docking_site)
}
\value{
number PSI
}
\description{
The number of read support a novel junction divided by the maximum number of read support the annotated junction that share a same splice site with the novel junction.
}
