% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapsedsQTL.R
\name{collapsedsQTL}
\alias{collapsedsQTL}
\title{Nominate csQTLs by matching SpliceAI predictions of the genetic variants to the provided splicing junctions}
\usage{
collapsedsQTL(
  junc_anno_file,
  paired_data,
  gdb_path,
  cohort_name,
  rna_meta = NULL,
  wgs_meta = NULL,
  max_allele_freq = 0.05,
  reference = "Default",
  spliceai_prediction = list(SpliceAI = c("spliceaiDS_AG", "spliceaiDS_AL",
    "spliceaiDS_DG", "spliceaiDS_DL")),
  min_spliceai_score = 0.2,
  spliceai_default_reference = FALSE,
  max_skip_exon = 2,
  blacklist = system.file(paste("extdata", "Blacklist", "hg38-blacklist-nochr.v2.bed.gz",
    sep = .Platform$file.sep), package = "SpliPath"),
  output_prefix = ""
)
}
\arguments{
\item{junc_anno_file}{string or data.frame The output file of annotateJunc function, which contain junction annotations.}

\item{paired_data}{logical Find collapsed sQTL in paired data mode "TRUE" or unpaired data mode "FALSE". If TRUE, arguments rna_meta and wgs_meta are required. The function matches the paired samples by 'SubjectID' in the metadata.}

\item{gdb_path}{string The path to GDB database that store the genotype data from whole genome sequencing.}

\item{cohort_name}{string The table name in GDB database that contain sample metadata}

\item{rna_meta}{string RNAseq sample metadata file. The file should contain at least "SampleID", "SubjectID", "Group, "Tissue" columns.}

\item{wgs_meta}{string WGS sample metadata file. The file should contain at least "SampleID", "SubjectID", "Group columns.}

\item{max_allele_freq}{numeric Maximum allele frequency to nominate csQTL candidate.}

\item{reference}{string The directory to genomic annotation files, which can be generated by prepareGenoRef function. Default to pre-made ensembl annotation version GRCh38.98.}

\item{spliceai_prediction}{list The names of elements in the list is the prediction tool, it should be in GDB database tables. The items are the scoring fields in the GDB database prediction table. The maximum of the given scores will be used as the prediction of the corresponding tool for each variants. E.g. The SpliceAI score of a variant will be the maximum of spliceaiDS_AG, "pliceaiDS_AL, spliceaiDS_DG, and spliceaiDS_DL. Default: list(SpliceAI = c("spliceaiDS_AG", "spliceaiDS_AL", "spliceaiDS_DG", "spliceaiDS_DL")).}

\item{spliceai_default_reference}{logical Set to TRUE if the SpliceAI predictions are pre-computed or SpliceAI default annotation were used when running SpliceAI (-A GRCh38, or -A GRCh37). Default to FALSE, indicating the annotation GTF are the same for junction annotation and SpliceAI prediction.}

\item{max_skip_exon}{integer If a novel junction skips more exons than the provided number, it will be considered as "not match" to prediction. Default to 2.}

\item{blacklist}{string or NULL Optional. A genomic region blacklist. If provided, the regions will be removed.  Default: system.file("extdata/Reference/blacklist/hg38-blacklist-nochr.v2.bed.gz", package="SpliPath").}

\item{output_prefix}{string The mapping results will be written in output_prefix_map.txt.gz}

\item{min_splice_score}{list The min splice score of each prediction tools to nominate ultra rare sQTL candidate. A min score for SpliceAI must be in the list. Default: list(SpliceAI = 0.2, dbscSNV = 0.7).The predictions of splice sites that score higher than the minimum SpliceAI score will be matched with observed junctions.}
}
\description{
Nominate csQTLs by matching SpliceAI predictions of the genetic variants to the provided splicing junctions
}
