% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapsedsQTL_pangolin.R
\name{collapsedsQTL_pangolin}
\alias{collapsedsQTL_pangolin}
\title{Nominate csQTLs by matching Pangolin predictions of the genetic variants to the provided splicing junctions}
\usage{
collapsedsQTL_pangolin(
  junc_anno_file,
  paired_data,
  gdb_path,
  cohort_name,
  rna_meta = NULL,
  wgs_meta = NULL,
  max_allele_freq = 0.05,
  reference = "Default",
  pangolin_prediction = list(pangolin = c("score_increase", "score_decrease")),
  min_pangolin_score = 0.2,
  max_skip_exon = 2,
  blacklist = system.file(paste("extdata", "Blacklist", "hg38-blacklist-nochr.v2.bed.gz",
    sep = .Platform$file.sep), package = "SpliPath"),
  output_prefix = ""
)
}
\arguments{
\item{junc_anno_file}{string or data.frame The output file of annotateJunc function, which contain junction annotations.}

\item{paired_data}{logical Find collapsed sQTL in paired data mode "TRUE" or unpaired data mode "FALSE". If TRUE, arguments rna_meta and wgs_meta are required. The function matches the paired samples by 'SubjectID' in the metadata.}

\item{gdb_path}{string The path to GDB database that store the genotype data from whole genome sequencing.}

\item{cohort_name}{string The table name in GDB database that contain sample metadata}

\item{rna_meta}{string RNAseq sample metadata file. The file should contain at least "SampleID", "SubjectID", "Group, "Tissue" columns.}

\item{wgs_meta}{string WGS sample metadata file. The file should contain at least "SampleID", "SubjectID", "Group columns.}

\item{max_allele_freq}{numeric Maximum allele frequency to nominate csQTL candidate.}

\item{reference}{string The directory to genomic annotation files, which can be generated by prepareGenoRef function. Default to pre-made ensembl annotation version GRCh38.98.}

\item{pangolin_prediction}{list The names of elements in the list is the prediction tool, it should be in GDB database tables. The items are the scoring fields in the GDB database prediction table. The maximum of the given scores will be used as the prediction of the corresponding tool for each variants. E.g. The Pangolin score of a variant will be the maximum of absolute "score_increase", "score_decrease". Default: list(pangolin = c("score_increase", "score_decrease")).}

\item{min_pangolin_score}{numeric The min splice score of the prediction tools to nominate rare sQTL candidate. The predictions of splice sites that score higher than the min score will be matched with observed junctions.}

\item{max_skip_exon}{integer If a novel junction skips more exons than the provided number, it will be considered as "not match" to prediction. Default to 2.}

\item{blacklist}{string or NULL Optional. A genomic region blacklist. If provided, the regions will be removed.  Default: system.file("extdata/Reference/blacklist/hg38-blacklist-nochr.v2.bed.gz", package="SpliPath").}

\item{output_prefix}{string The mapping results will be written in output_prefix_map.txt.gz}
}
\description{
Nominate csQTLs by matching Pangolin predictions of the genetic variants to the provided splicing junctions
}
