% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countBackGoundJunc.R
\name{countBackGroundJunc}
\alias{countBackGroundJunc}
\title{Count the number of reads support the background junctions of the novel junctions}
\usage{
countBackGroundJunc(
  junc_file,
  junc_anno,
  output_prefix,
  annotation = c("in.ens38", "in.snaptron.gtex")
)
}
\arguments{
\item{junc_file}{string or data.frame The output file of mergeJuncData function, which contain junction-sample read counts table. The first four columns should be the coordinates of each junction: "chr", "start", "end", "strand".}

\item{junc_anno}{string or data.frame The output file of annotateJunc function, which contain junction annotations.}

\item{output_prefix}{string Prefix of output PSI file. The junction-sample PSI table will be written in a prefix_norm.txt.gz file.}

\item{annotation}{vector Annotation sources (should in colnames in junc_anno) that were combined to determine novel junction. A junction not annotated in any given source is treated as novel junction. Default: c("in.ensembl", "in.snaptron")}
}
\value{
data.frame Seven-column dataframe. Contain coordinates, mapped gene, and splicing event of each novel junction.
}
\description{
Count the number of reads support the background junctions of the novel junctions
}
