% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNovelJuncExp.R
\name{getNovelJuncExp}
\alias{getNovelJuncExp}
\title{Quantify the expression of novel junctions based the expression of the background junctions}
\usage{
getNovelJuncExp(
  junc_file,
  junc_anno,
  annotation = c("in.ensembl", "in.snaptron"),
  rna_meta,
  gene_expression_file_prefix,
  output_prefix
)
}
\arguments{
\item{junc_file}{string or data.frame The output file of mergeJuncData function, which contain junction-sample read counts table. The first four columns should be the coordinates of each junction: "chr", "start", "end", "strand".}

\item{junc_anno}{string or data.frame The output file of annotateJunc function, which contain junction annotations.}

\item{annotation}{vector Annotation sources (should in colnames in junc_anno) that were combined to determine novel junction. A junction not annotated in any given source is treated as novel junction. Default: c("in.ensembl", "in.snaptron")}

\item{rna_meta}{string or data.frame The meta data for RNAseq samples. It should contain at least 4 columns: SubjectID, SampleID, Group, Tissue.}

\item{gene_expression_file_prefix}{string The file prefix of the corrected gene expression data, generated by function correctGeneExp.}

\item{output_prefix}{string Prefix of output PSI file. The junction-sample PSI table will be written in a prefix_norm.txt.gz file.}
}
\value{
data.frame Seven-column dataframe. Contain junction coordinates, mapped gene, and splicing event of each novel junction.
}
\description{
Quantify the expression of novel junctions based the expression of the background junctions
}
