% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juncVariantRegion.R
\name{juncVariantRegion}
\alias{juncVariantRegion}
\title{Find the variants mapping regions where the possible cis-acting splice-altering variants lie in}
\usage{
juncVariantRegion(novel_junc, output_prefix, reference = "Default")
}
\arguments{
\item{novel_junc}{dataframe The returned dataframe of callJuncPSI function. Contain coordinates, mapped gene, and splicing event of each novel junction.}

\item{output_prefix}{string Prefix of output file. The variant mapping region table will be written in a prefix_variant_region.txt.gz file.}

\item{reference}{string The directory to genomic annotation files, which can be generated by prepareGenoRef function. Default to pre-made ensembl annotation version GRCh38.98.}
}
\value{
data.frame Each row is a novel junction and the coordinates of region (column "region.start" and "region.end") where the possible cis-acting splice-altering variants lie in.
}
\description{
The variants mapping region of a novel junction is the disrupted intron and flanking exons.
}
