% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juncVariantRegion_.R
\name{juncVariantRegion_}
\alias{juncVariantRegion_}
\title{Find the intron disrupted by the unannotated junction and the two flanking exons}
\usage{
juncVariantRegion_(junc, novel_junc, exon)
}
\description{
Find the intron disrupted by the unannotated junction and the two flanking exons
}
