% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapVariantJunc.R
\name{mapVariantJunc}
\alias{mapVariantJunc}
\title{Map potential splice-altering DNA variants to novel junctions in tissues}
\usage{
mapVariantJunc(
  junc_var_region,
  psi_file,
  count_file,
  gdb_path,
  cohort_name,
  rna_meta,
  wgs_meta,
  tissues,
  allele_freq = "gdb",
  splice_prediction = list(SpliceAI = c("spliceaiDS_AG", "spliceaiDS_AL",
    "spliceaiDS_DG", "spliceaiDS_DL"), dbscSNV = c("rf_score", "ada_score")),
  output_prefix = ""
)
}
\arguments{
\item{junc_var_region}{string or data.frame The output file or returned data.frame of juncVariantRegion function. The file contains the variants mapping region of each novel junction.}

\item{psi_file}{string The output file of calJuncPSI function, which contain junction-sample PSI table.}

\item{count_file}{string The output file of mergeJuncData funciton, which contain junction-sample read count table.}

\item{gdb_path}{string The path to GDB database that store the genotype data from whole genome sequencing.}

\item{cohort_name}{string The table name in GDB database that contain sample metadata}

\item{rna_meta}{string RNAseq sample metadata file. The file should contain at least "SampleID", "SubjectID", "Tissue" columns.}

\item{wgs_meta}{string WGS sample metadata file. The file should contain at least "SampleID", "SubjectID" columns.}

\item{tissues}{vector potential splice-altering DNA variants to novel junctions in given tissues.}

\item{allele_freq}{list or string Allele frequency of the variants. Default: "gdb", the allele frequency among the whole samples in the provided GDB. Or use other allele frequency annotations of the variants recorded in the GDB tables, e.g. list(gdb_table = "gnomAD", af_field = "AF"). The gdb_table and af_field are the name of variant annotation table and field where the allele frequency is in the GDB.}

\item{splice_prediction}{list The splice-altering prediction to annotate the DNA variants. Default: list(SpliceAI = c("spliceaiDS_AG", "spliceaiDS_AL", "spliceaiDS_DG", "spliceaiDS_DL"), dbscSNV = c("rf_score", "ada_score")). 
The names of elements in the list are prediction tools, they should be in GDB database tables. The items are the scoring fields in the GDB database prediction tables. The maximum of the given scores will be used as the prediction of the corresponding tool for each variants. E.g. The SpliceAI score of a variant will be the maximum of spliceaiDS_AG, "pliceaiDS_AL, spliceaiDS_DG, and spliceaiDS_DL.}

\item{output_prefix}{string The mapping results will be written in output_prefix_map.txt.gz}
}
\value{
data.frame One row in the the dataframe is an annotated DNA variant that mapped to a novel junction in a subject's tissues. The junction PSI and read count in tissues are in columns "PSI_tissue" and "Read_tissue".
}
\description{
For each individual, map the DNA variants that of splice-altering potential to a novel junction in this individual's tissues.
}
