% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchPredNObs.R
\name{matchPredNObs}
\alias{matchPredNObs}
\title{Match SpliceAI prediction of splice sites gain or loss with observed junctions}
\usage{
matchPredNObs(
  gdb_path,
  cohort_name,
  var2junc,
  min_SpliceAI_score,
  max_skip_exon = 2,
  reference = "Default",
  spliceai_default_reference = FALSE
)
}
\arguments{
\item{gdb_path}{string The path to GDB database that store the genotype data from whole genome sequencing. The GDB database should contain SpliceAI prediction table.}

\item{cohort_name}{string The table name in GDB database that contain sample metadata}

\item{var2junc}{string or data.frame One row in the the dataframe is an annotated DNA variant that mapped to a novel junction in a subject's tissues.}

\item{min_SpliceAI_score}{float Minimum SpliceAI delta score. The predictions of splice sites that score higher than the minimum SpliceAI delta score will be matched with observed junctions.}

\item{max_skip_exon}{integer If a novel junction skips more exons than the provided number, it will be considered as "not match" to prediction. Default to 2.}

\item{reference}{string The directory to genomic annotation files, which can be generated by prepareGenoRef function. Default to pre-made ensembl annotation version GRCh38.98.}

\item{spliceai_default_reference}{logical Set to TRUE if the SpliceAI predictions are pre-computed or SpliceAI default annotation were used when running SpliceAI (-A GRCh38, or -A GRCh37). Default to FALSE, indicating the annotation GTF are the same for junction annotation and SpliceAI prediction.}
}
\value{
data.frame Column "match_by_threshold" shows whether the prediction of variants matched the corresponding novel junction. "NA" in the column means either the prediction score is under given threshold, or the prediction is a gain of annotated site or a loss of novel site which is out the scope of this function. 
The "SG", "SL", "EG", and "EL" in the column names means "start gain", "start loss", "end gain", and "end loss".
}
\description{
Match SpliceAI prediction of splice sites gain or loss with observed junctions
}
