% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchPredNObs_pangolin.R
\name{matchPredNObs_pangolin}
\alias{matchPredNObs_pangolin}
\title{Match Pangolin prediction of splice sites gain or loss with observed junctions}
\usage{
matchPredNObs_pangolin(
  gdb_path,
  cohort_name,
  var2junc,
  pangolin_table_name,
  min_pangolin_score,
  max_skip_exon = 2,
  reference = "Default"
)
}
\arguments{
\item{gdb_path}{string The path to GDB database that store the genotype data from whole genome sequencing. The GDB database should contain Pangolin prediction table.}

\item{cohort_name}{string The table name in GDB database that contain sample metadata}

\item{var2junc}{string or data.frame One row in the the dataframe is an annotated DNA variant that mapped to a novel junction in a subject's tissues.}

\item{pangolin_table_name}{string The table of Pangolin predictions in the GDB database}

\item{min_pangolin_score}{float Minimum Pangolin score. The predictions of splice sites that score higher than the minimum score will be matched with observed junctions.}

\item{max_skip_exon}{integer If a novel junction skips more exons than the provided number, it will be considered as "not match" to prediction. Default to 2.}

\item{reference}{string The directory to genomic annotation files, which can be generated by prepareGenoRef function. Default to pre-made ensembl annotation version GRCh38.98.}
}
\value{
data.frame Column "match_by_threshold" shows whether the prediction of variants matched the corresponding novel junction. "NA" in the column means either the prediction score is under given threshold, or the prediction is a gain of annotated site or a loss of novel site which is out the scope of this function.
}
\description{
Match Pangolin prediction of splice sites gain or loss with observed junctions
}
