% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nominateHotspot.R
\name{nominateHotspot}
\alias{nominateHotspot}
\title{Nominate hotspots of splice-altering variants from variants that mapped to novel junctions}
\usage{
nominateHotspot(
  tissue_var2junc,
  min_psi,
  min_read,
  min_read_include,
  tissues,
  min_nr_tissue,
  min_splice_pred = list(SpliceAI = 0.2, dbscSNV = 0.7),
  max_allele_freq = 0.05,
  splicai_pred_match = NULL,
  blacklist = system.file("data/Reference/Blacklist/hg38-blacklist-nochr.v2.bed.gz",
    package = "SpliPath"),
  output_prefix = ""
)
}
\arguments{
\item{tissue_var2junc}{string or data.frame The output The output file or data.frame of mapVariantJunc function. One row in the the dataframe is an annotated DNA variant that mapped to a novel junction in a subject's tissues.}

\item{min_psi}{number The minimum PSI required for observed novel junction.}

\item{min_read}{number The minimum number of reads required to support the observed novel junction.}

\item{min_read_include}{number Also include junctions supported by more than min_read_include reads even of their PSI are under min_psi.}

\item{tissues}{vector The names of tissues.}

\item{min_nr_tissue}{number Require novel junction observation concordant across min_nr_tissue tissues.}

\item{min_splice_pred}{list The minimum prediction score of prediciton tools. Default: list(SpliceAI = 0.2, dbscSNV = 0.7).}

\item{max_allele_freq}{string The maximum allele frequency.}

\item{splicai_pred_match}{string or data.frame Optional. The output file or dataframe of matchPredNObs function, which shows whether the SpliceAI prediction of a variant match an observed novel junction. If provided, information of whether predictions match observations will be add in the output file.}

\item{blacklist}{string or NULL Optional. A genomic region blacklist. If provided, the regions will be removed.  Default: system.file("extdata/Reference/blacklist/hg38-blacklist-nochr.v2.bed.gz", package="SpliPath").}

\item{output_prefix}{string The nominated mutation hotspot will be written in output_prefix_hotspot.txt.gz. The "Hotspot" column shows whether the variants and novel junctions are passed the given thresholds and nominated as mutation hotspots.}
}
\description{
Nominate hotspots of splice-altering variants from variants that mapped to novel junctions
}
