% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normCount.R
\name{normCount}
\alias{normCount}
\title{Calculate PSI of novel junctions within each gene}
\usage{
normCount(junc_count)
}
\value{
data.frame
}
\description{
A function to connect calJuncPSI function and normCount_ function. If there are only novel junctions in a gene, the PSI will not be calculated (set to 0 by calJuncPSI function).
}
