% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normCount_.R
\name{normCount_}
\alias{normCount_}
\title{Calculate PSI of different splicing events in a gene}
\usage{
normCount_(gene_junc)
}
\value{
data.frame
}
\description{
The function to connect normCount function and calNormValue function.
}
