% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareGenoRef.R
\name{prepareGenoRef}
\alias{prepareGenoRef}
\title{Convert genome annotation GTF file to reference gene, intron, exon annotation files}
\usage{
prepareGenoRef(gtf_file, intron_bed, output_dir)
}
\arguments{
\item{gtf_file}{string Ensembl genomic annotation GTF file}

\item{intron_bed}{string BED file contain coordinates (chromosome, start position, end position, strand), transcript id, and gene id information of introns. This file can be generated by GTFtools (https://www.genemine.org/gtftools.php): python gtftools.py -i introns.bed ensembl.gtf}

\item{output_dir}{string The output annotation files will be written into the provided directory. The four output files are "Gene.bed": coordinates of genes; "Exon_proteincoding.bed": exons in protein coding transcripts; "Intron_proteincoding.bed": introns in protein coding transcripts; "Intron_noncoding.bed" introns in noncoding transcripts.}
}
\description{
Convert genome annotation GTF file to reference gene, intron, exon annotation files
}
