% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDataBrowser.R
\name{runDataBroswer}
\alias{runDataBroswer}
\title{Run SpliPath data browser}
\usage{
runDataBroswer(
  paired_data,
  gdb_path,
  data_dir,
  rna_meta = NULL,
  dna_meta = NULL,
  reference = "Default"
)
}
\arguments{
\item{gdb_path}{string Path to GDB file that store genomic variants data of the DNA sequencing samples.}

\item{data_dir}{string Path to junction data prepared for SpliPath data browser by browserData function. It should be the same with the "output_dir" argument of browserData function.}

\item{rna_meta}{string Path to RNAseq sample metadata file. There should be at least "SampleID", "SubjecID", "Tissue", "Group" columns in the file.}

\item{dna_meta}{string Path to DNA sequencing sample metadata file. There should be at least "SampleID", "SubjecID" columns in the file.}

\item{reference}{string The directory to genomic annotation files, which can be generated by prepareGenoRef function. Default to pre-made ensembl annotation version GRCh38.98.}
}
\description{
Run SpliPath data browser
}
