% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissueVar2Junc.R
\name{tissueVar2Junc}
\alias{tissueVar2Junc}
\title{Map potential splice-altering DNA variants to novel junctions in one tissue}
\usage{
tissueVar2Junc(rna_meta, tissue, var2junc, psi, read_count, geno, var_pred)
}
\value{
data.frame One row in the the dataframe is an annotated DNA variant that mapped to a novel junction in a subject's tissues. The junction PSI and read count in the are in columns "PSI_tissue" and "Read_tissue".
}
\description{
For each individual, map the DNA variants that of splice-altering potential to a novel junction in this individual's tissue.
}
