% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateJunc.R
\name{annotateJunc}
\alias{annotateJunc}
\title{Annotate junctions extracted from RNAseq data}
\usage{
annotateJunc(
  rna_meta,
  chrom,
  output_prefix,
  reference = "Default",
  addition_annotation = list(snaptron = system.file(paste("extdata", "Reference",
    "snaptron_gtex_anno.txt", sep = .Platform$file.sep), package = "SpliPath"))
)
}
\arguments{
\item{rna_meta}{string RNAseq sample metadata file. The file should contain at least "SampleID", "SubjectID", "Group", "Tissue", and "Path" columns}

\item{chrom}{string The junctions in this chromosome will be merged among samples.}

\item{output_prefix}{string Prefix of output file. The annotations of the junctions will be written in a prefix_anno.txt.gz file.}

\item{reference}{string The directory to genomic annotation files, which can be generated by prepareGenoRef function. Default to pre-made ensembl annotation version GRCh38.98.}

\item{addition_annotation}{list or NULL Optional. Name and file name of additional junction annotation source. Default: list(snaptron = system.file("extdata/Reference/snaptron_gtex_anno.txt", package = "SpliPath")). Whether the input junctions are the given junction list will be shown in the output file (in column "in.annotation_name").}
}
\description{
Annotate junctions extracted from RNAseq data
}
