% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browserData.R
\name{browserData}
\alias{browserData}
\title{Prepare data files for SpliPath data browser}
\usage{
browserData(
  junc_count,
  junc_anno,
  tissues_leafcutter_pvals_file,
  var2junc,
  min_read = 5,
  output_dir = "."
)
}
\arguments{
\item{junc_count}{vector A list of output files of mergeJuncData funciton, which contain junction-sample read count table. The input file name list in the first four arguments should be matched.}

\item{junc_anno}{vector A list of output files of annotateJunc function, which contain junction annotations.}

\item{tissues_leafcutter_pvals_file}{list The file paths to the LeafCutter analysis P values of each tissue type. Each element is a LeafCutter file path named under the tissue type. The tissues should be the same with those in the RNAseq sample metadata file.}

\item{var2junc}{string or data.frame The output file or data.frame of mapVariantJunc function. One row in the the dataframe is an annotated DNA variant that mapped to a novel junction in a subject's tissues.}

\item{output_dir}{string The path to data output directory.}
}
\description{
Separate splice junction tables and sRV tables per gene. The separated files will be written in SpliPath_Browser/data/junction and SpliPath_Browser/data/sRV_candidate
}
