% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browserData.R
\name{browserData}
\alias{browserData}
\title{Prepare data files for SpliPath data browser}
\usage{
browserData(
  junc_count,
  junc_psi,
  junc_anno,
  hotspot,
  min_read = 5,
  output_dir = "."
)
}
\arguments{
\item{junc_count}{vector A list of output files of mergeJuncData funciton, which contain junction-sample read count table. The input file name list in the first four arguments should be matched.}

\item{junc_psi}{vector A list of output files of calJuncPSI function, which contain junction-sample PSI table.}

\item{junc_anno}{vector A list of output files of annotateJunc function, which contain junction annotations.}

\item{hotspot}{vector A list of output files of nominateHotspot function.}

\item{output_dir}{string The path to data output directory.}
}
\description{
Separate splice junction tables and sRV tables per gene. The separated files will be written in SpliPath_Browser/data/junction and SpliPath_Browser/data/sRV_candidate
}
