% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapVariantJunc.R
\name{mapVariantJunc}
\alias{mapVariantJunc}
\title{Map potential splice-altering DNA variants to novel junctions in tissues}
\usage{
mapVariantJunc(
  junc_anno_file,
  read_count_file,
  tissues_leafcutter_pvals_file,
  gdb_path,
  cohort_name,
  rna_meta,
  wgs_meta,
  reference = "Default",
  as_annotated = c("in.ensembl", "in.snaptron"),
  max_LeafCutter_Pval = 0.05,
  min_nr_tissue = 1,
  source_allele_freq = "cohort",
  max_allele_freq = 0.05,
  splice_prediction = list(SpliceAI = c("spliceaiDS_AG", "spliceaiDS_AL",
    "spliceaiDS_DG", "spliceaiDS_DL"), dbscSNV = c("rf_score", "ada_score")),
  min_splice_score = list(SpliceAI = 0.2, dbscSNV = 0.7),
  SpliceAI_default_reference = FALSE,
  max_skip_exon = 2,
  blacklist = system.file(paste("extdata", "Blacklist", "hg38-blacklist-nochr.v2.bed.gz",
    sep = .Platform$file.sep), package = "SpliPath"),
  output_prefix = ""
)
}
\arguments{
\item{junc_anno_file}{string or data.frame The output file of annotateJunc function, which contain junction annotations.}

\item{read_count_file}{string The output file of annotateJunc funciton, which contain junction-sample read count table.}

\item{tissues_leafcutter_pvals_file}{list The file paths to the LeafCutter analysis P values of each tissue type. Each element is a LeafCutter file path named under the tissue type. The tissues should be the same with those in the RNAseq sample metadata file.}

\item{gdb_path}{string The path to GDB database that store the genotype data from whole genome sequencing.}

\item{cohort_name}{string The table name in GDB database that contain sample metadata}

\item{rna_meta}{string RNAseq sample metadata file. The file should contain at least "SampleID", "SubjectID", "Group, "Tissue" columns.}

\item{wgs_meta}{string WGS sample metadata file. The file should contain at least "SampleID", "SubjectID", "Group columns.}

\item{reference}{string The directory to genomic annotation files, which can be generated by prepareGenoRef function. Default to pre-made ensembl annotation version GRCh38.98.}

\item{as_annotated}{vector Annotation sources (should be in colnames in junc_anno) that were combined to determine novel junction. A junction not annotated in any given source is treated as novel junction. Default: c("in.ensembl", "in.snaptron")}

\item{max_LeafCutter_Pval}{numeric Maximum LeafCutter P value of a novel junction to nominate ultra rare sQTL candidate.}

\item{min_nr_tissue}{numeric Require novel junction LeafCutter P value under the provided threshold in at least min_nr_tissue tissues.}

\item{source_allele_freq}{list or string Source of the variants allele frequency. Default: "cohort", the allele frequency in the cohort provided by the argument "cohort_name". Or use other allele frequency annotations of the variants recorded in GDB tables, e.g. list(gdb_table = "gnomAD", af_field = "AF"). The gdb_table and af_field are the name of variant annotation table and field where the allele frequency is in the GDB.}

\item{max_allele_freq}{numeric Maximum allele frequency to nominate ultra rare sQTL candidate.}

\item{splice_prediction}{list The splice-altering prediction to annotate the DNA variants. Default: list(SpliceAI = c("spliceaiDS_AG", "spliceaiDS_AL", "spliceaiDS_DG", "spliceaiDS_DL"), dbscSNV = c("rf_score", "ada_score")). The names of elements in the list are prediction tools, they should be in GDB database tables. The items are the scoring fields in the GDB database prediction tables. The maximum of the given scores will be used as the prediction of the corresponding tool for each variants. E.g. The SpliceAI score of a variant will be the maximum of spliceaiDS_AG, "pliceaiDS_AL, spliceaiDS_DG, and spliceaiDS_DL.}

\item{min_splice_score}{list The min splice score of each prediction tools to nominate ultra rare sQTL candidate. A min score for SpliceAI must be in the list. Default: list(SpliceAI = 0.2, dbscSNV = 0.7).The predictions of splice sites that score higher than the minimum SpliceAI score will be matched with observed junctions.}

\item{SpliceAI_default_reference}{logical Set to TRUE if the SpliceAI predictions are pre-computed or SpliceAI default annotation were used when running SpliceAI (-A GRCh38, or -A GRCh37). Default to FALSE, indicating the annotation GTF are the same for junction annotation and SpliceAI prediction.}

\item{max_skip_exon}{integer If a novel junction skips more exons than the provided number, it will be considered as "not match" to prediction. Default to 2.}

\item{blacklist}{string or NULL Optional. A genomic region blacklist. If provided, the regions will be removed.  Default: system.file("extdata/Reference/blacklist/hg38-blacklist-nochr.v2.bed.gz", package="SpliPath").}

\item{output_prefix}{string The mapping results will be written in output_prefix_map.txt.gz}
}
\value{
data.frame One row in the the data.frame is an annotated DNA variant that mapped to a novel junction in a subject's tissues. The junction LeafCutter outlier P value and read count in tissues are in columns "LeafCutter_pVal_tissue" and "Read_tissue".
}
\description{
For each individual, map the DNA variants that of splice-altering potential to a novel junction in this individual's tissues.
}
