% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchPredNObs.R
\name{matchPredNObs}
\alias{matchPredNObs}
\title{Match SpliceAI prediction of splice sites gain or loss with observed junctions}
\usage{
matchPredNObs(gdb_path, var2junc, min_spliceai_score, reference = "Default")
}
\arguments{
\item{gdb_path}{string The path to GDB database that store the genotype data from whole genome sequencing. The GDB database should contain SpliceAI prediction table.}

\item{var2junc}{string or data.frame The output file or data.frame of mapVariantJunc function. One row in the the dataframe is an annotated DNA variant that mapped to a novel junction in a subject's tissues.}

\item{min_spliceai_score}{number Minimum SpliceAI delta score. The predictions of splice sites that score higher than the minimum SpliceAI delta score will be matched with observed junctions.}

\item{reference}{string The directory to genomic annotation files, which can be generated by prepareGenoRef function. Default to pre-made ensembl annotation version GRCh38.98.}
}
\value{
data.frame Column "match_by_threshold" shows whether the prediction of variants matched the corresponding novel junction. "NA" in the column means either the prediction score is under given threshold, or the prediction is a gain of annotated site or a loss of novel site which is out the scope of this function. 
The "SG", "SL", "EG", and "EL" in the column names means "start gain", "start loss", "end gain", and "end loss".
}
\description{
Match SpliceAI prediction of splice sites gain or loss with observed junctions
}
